PR #4603.

+++ awscli/testutils.py
@@ -203,7 +203,7 @@ def random_chars(num_chars):
     return binascii.hexlify(os.urandom(int(num_chars / 2))).decode('ascii')
 
 
-def random_bucket_name(prefix='awscli-s3integ-', num_random=10):
+def random_bucket_name(prefix='awscli-s3integ-', num_random=15):
     """Generate a random S3 bucket name.
 
     :param prefix: A prefix to use in the bucket name.  Useful
@@ -509,7 +509,8 @@ class FileCreator(object):
         self.rootdir = tempfile.mkdtemp()
 
     def remove_all(self):
-        shutil.rmtree(self.rootdir)
+        if os.path.exists(self.rootdir):
+            shutil.rmtree(self.rootdir)
 
     def create_file(self, filename, contents, mtime=None, mode='w'):
         """Creates a file in a tmpdir
@@ -715,6 +716,12 @@ class BaseS3CLICommand(unittest.TestCase):
     and more streamlined.
 
     """
+    _PUT_HEAD_SHARED_EXTRAS = [
+        'SSECustomerAlgorithm',
+        'SSECustomerKey',
+        'SSECustomerKeyMD5',
+        'RequestPayer',
+    ]
 
     def setUp(self):
         self.files = FileCreator()
@@ -746,6 +753,7 @@ class BaseS3CLICommand(unittest.TestCase):
         return client
 
     def assert_key_contents_equal(self, bucket, key, expected_contents):
+        self.wait_until_key_exists(bucket, key)
         if isinstance(expected_contents, six.BytesIO):
             expected_contents = expected_contents.getvalue().decode('utf-8')
         actual_contents = self.get_key_contents(bucket, key)
@@ -777,6 +785,18 @@ class BaseS3CLICommand(unittest.TestCase):
             call_args.update(extra_args)
         response = client.put_object(**call_args)
         self.addCleanup(self.delete_key, bucket_name, key_name)
+        extra_head_params = {}
+        if extra_args:
+            extra_head_params = dict(
+                (k, v) for (k, v) in extra_args.items()
+                if k in self._PUT_HEAD_SHARED_EXTRAS
+            )
+        self.wait_until_key_exists(
+            bucket_name,
+            key_name,
+            extra_params=extra_head_params,
+        )
+        return response
 
     def delete_bucket(self, bucket_name, attempts=5, delay=5):
         self.remove_all_objects(bucket_name)
@@ -816,6 +836,7 @@ class BaseS3CLICommand(unittest.TestCase):
         response = client.delete_object(Bucket=bucket_name, Key=key_name)
 
     def get_key_contents(self, bucket_name, key_name):
+        self.wait_until_key_exists(bucket_name, key_name)
         client = self.create_client_for_bucket(bucket_name)
         response = client.get_object(Bucket=bucket_name, Key=key_name)
         return response['Body'].read().decode('utf-8')
